majorly rustify and simplify the incoming decryption pipeline
