+++ src/sql.rs
@@ -777,7 +777,7 @@ fn open(
                         if let Some(ref mut peerstate) = Peerstate::from_addr(context, sql, &addr?)
                         {
                             peerstate.recalc_fingerprint();
-                            peerstate.save_to_db(sql, false);
+                            peerstate.save_to_db(sql, false).unwrap();
                         }
                     }
                     Ok(())
